refactor schedule structure
