+++ megatron/spiral/spipe_schedule.py
@@ -1,0 +1,677 @@
+import contextlib
+import warnings
+import nvtx
+import sys
+from typing import Callable, Iterator, List, Optional, Union
+from collections import deque
+from queue import Queue
+import threading
+
+import torch
+from torch.nn.parallel.distributed import DistributedDataParallel as torchDDP
+
+from megatron import get_args, get_timers
+from megatron.core import mpu
+from megatron.core.enums import ModelType
+from megatron.core.utils import get_model_type, get_attr_wrapped_model
+from megatron.core.pipeline_parallel import forward_step, backward_step
+from megatron.spiral.initialize import get_thunder_cuda_manager
+from megatron.spiral.debug import spiral_print
+import megatron.spiral.p2p_communication as spiral_p2p
+from megatron.spiral.init_context import SpiralParamStatus, set_module_spiral_status
+from megatron.spiral.utils import is_spiral_param
+from megatron.spiral.generic import ContextManagers
+import megatron.spiral.build_state as sbs
+from megatron.spiral.ckpt_communication import comm_input_ckpt
+from megatron.spiral.ckpt_schedule import CkptSendRecvType, CkptSendRecvSchedule, CkptSendRecvOp
+
+
+# Types
+Shape = Union[List[int], torch.Size]
+
+# Constants
+_DEBUG_SCHEDULE = True
+
+
+def spipe_schedule(
+    *,
+    forward_step_func,
+    data_iterator: Union[Iterator, List[Iterator]],
+    model: Union[torch.nn.Module, List[torch.nn.Module]],
+    num_microbatches: int,
+    dtype: torch.dtype,
+    tensor_shape: Shape,
+    decoder_seq_length: Optional[int] = None,
+    grad_scaler: Callable = None,
+    sequence_parallel: bool = False,
+    overlap_p2p_comm: bool = False,
+    batch_p2p_comm: bool = False,
+    forward_only: bool = False,
+    timers: Callable = None,
+    collect_non_loss_data: bool = False,
+    enable_autocast: bool = False,
+    deallocate_pipeline_outputs: bool = False,
+    no_sync_func: Optional[Callable] = None,
+    grad_sync_func: Optional[Callable] = None,
+    param_sync_func: Optional[Callable] = None,
+    **kwargs,
+):
+    """Run sprial schedule, with communication between pipeline stages as needed.
+
+    Returns dictionary with losses if the last stage, empty dict otherwise."""
+
+    assert isinstance(
+        model, list
+    ), "SpiralPipe expected model chunking by stage"
+    assert isinstance(
+        data_iterator, list
+    ), "SpiralPipe expected each model chunk to have a data iterator"
+
+    # TODO (SpiralPipe) disable async grad reductions?
+    if no_sync_func is None and all(isinstance(chunk, torchDDP) for chunk in model):
+
+        def multi_no_sync():
+            stack = contextlib.ExitStack()
+            for chunk in model:
+                stack.enter_context(chunk.no_sync())
+            return stack
+
+        no_sync_func = multi_no_sync
+    if no_sync_func is None:
+        no_sync_func = contextlib.nullcontext
+    no_sync_context = None
+
+    def disable_grad_sync():
+        """Disable asynchronous grad reductions"""
+        nonlocal no_sync_context
+        if no_sync_context is None:
+            no_sync_context = no_sync_func()
+            no_sync_context.__enter__()
+
+    def enable_grad_sync():
+        """Enable asynchronous grad reductions"""
+        nonlocal no_sync_context
+        if no_sync_context is not None:
+            no_sync_context.__exit__(None, None, None)
+            no_sync_context = None
+
+    disable_grad_sync()
+
+    if mpu.get_spiral_forward_virtual_rank() is not None:
+        warnings.warn(
+            "SpiralPipe forward virtual rank is not None on scheule entry. There may be a bug."
+        )
+    if mpu.get_spiral_backward_virtual_rank() is not None:
+        warnings.warn(
+            "SpiralPipe backward virtual rank is not None on scheule entry. There may be a bug."
+        )
+
+    if sequence_parallel:
+        seq_length, batch_size, hidden = tensor_shape
+        tensor_shape = (
+            seq_length // mpu.get_tensor_model_parallel_world_size(),
+            batch_size,
+            hidden,
+        )
+
+    model_type = get_model_type(model[0])
+    if model_type == ModelType.encoder_and_decoder:
+        raise RuntimeError("Spiral is not supported with an encoder and decoder model.")
+
+    if decoder_seq_length is not None and decoder_seq_length != tensor_shape[0]:
+        raise RuntimeError(
+            "Spiral is not supported with a different decoder sequence length."
+        )
+
+    offload_grad_after_bwd_stage = get_args().spiral_overlap_offload_grad
+    optimize_after_bwd_stage = False
+    if offload_grad_after_bwd_stage and get_args().spiral_stage_optimizer:
+        assert "spiral_stage_optimizer" in kwargs
+        assert "spiral_grad_scaler" in kwargs
+        assert "spiral_stage_optimizer_step_returns" in kwargs and isinstance(kwargs["spiral_stage_optimizer_step_returns"], deque)
+        optimize_after_bwd_stage = True
+        optimizer = kwargs["spiral_stage_optimizer"]
+        grad_scaler = kwargs["spiral_grad_scaler"]
+    if optimize_after_bwd_stage:
+        optimizer_threads = [] # (thread, queue, ev)
+
+    def _cleanup():
+        # cleanup checkpointed input tensors
+        for module in model:
+            empty_input_tensors: Callable = get_attr_wrapped_model(
+                module, "empty_input_tensors"
+            )
+            empty_input_tensors()
+
+    # Init input ckpt send recv schedule
+    ckpt_send_recv_schedule = None  # placeholder
+    if not forward_only:
+        ckpt_send_recv_schedule = CkptSendRecvSchedule(num_microbatches)
+
+    # Data structures for training
+    forward_data_store = []
+
+    # Data structures for delayed send
+    # Last pipeline worker will delay sending output tensors of non-last fwd stage
+    # Last pipeline worker will delay sending input tensor grads of non-first bwd stage
+    # TODO (SpiralPipe) Refactor for cleaner code
+    if mpu.get_pipeline_model_parallel_rank() == mpu.get_pipeline_model_parallel_world_size() - 1:
+        NUM_DELAY_OUTPUT_TENSORS = num_microbatches - mpu.get_pipeline_model_parallel_world_size()
+        NUM_DELAY_INPUT_TENSOR_GRADS = num_microbatches - mpu.get_pipeline_model_parallel_world_size()
+        curr_delay_output_tensors = 0
+        curr_delay_input_tensor_grads = 0
+        delayed_output_tensors = []
+        delayed_input_tensor_grads = []
+
+    # Event dictionaries. Key is the event tag, thus an event necessarily requires it.
+    prefetch_event_queries = {}
+    compute_event_queries = {}
+    offload_event_queries = {}
+    free_event_queries = {}
+
+    # Placeholders
+    recv_handles = None
+
+    """ Start training """
+
+    # nop pre-pipeline non-compute timesteps
+    __num_pre_pipeline_non_compute_ts = mpu.get_pipeline_model_parallel_rank()
+    for _ in range(__num_pre_pipeline_non_compute_ts):
+        if not forward_only:
+            next(ckpt_send_recv_schedule)
+
+    # prefetch 1st fwd stage
+    with torch.cuda.stream(get_thunder_cuda_manager().Stream("prefetch")):
+        model[0].spiral_fetch(non_blocking=True)
+        prefetch_f0 = get_thunder_cuda_manager().Event(
+            "prefetch",
+            "compute",
+            tag="prefetch:f0",
+            post_wait_fn=lambda: set_module_spiral_status(
+                model[0], SpiralParamStatus.GPU
+            ),
+        )
+        if get_thunder_cuda_manager().record_event(prefetch_f0) == -1:
+            raise RuntimeError("record_event failed")
+        prefetch_event_queries[prefetch_f0.tag] = prefetch_f0
+
+    # fwd
+    for fwd_stage_id in range(mpu.get_spiral_forward_virtual_size()):
+        if _DEBUG_SCHEDULE:
+            spiral_print(f"Start fwd stage {fwd_stage_id}")
+        mpu.set_spiral_forward_virtual_rank(fwd_stage_id)
+
+        assert (
+            hasattr(model[fwd_stage_id], "spiral_forward_stage_id")
+            and getattr(model[fwd_stage_id], "spiral_forward_stage_id") == fwd_stage_id
+        ), "Forward stage ID mismatch between virtual rank and model."
+
+        # prefetch next stage
+        with torch.cuda.stream(get_thunder_cuda_manager().Stream("prefetch")):
+            if not fwd_stage_id == 0:
+                if (
+                    get_thunder_cuda_manager().wait_event(free_event_queries.pop(f"free:f{fwd_stage_id - 1}"))
+                    == -1
+                ):
+                    raise RuntimeError("wait_event failed")
+
+            # skip prefetch if forward only and processing last fwd stage
+            if not (
+                forward_only
+                and fwd_stage_id == mpu.get_spiral_forward_virtual_size() - 1
+            ):
+                model[fwd_stage_id + 1].spiral_fetch(non_blocking=True)
+                tag = "prefetch:" + (
+                    f"f{fwd_stage_id + 1}"
+                    if fwd_stage_id + 1 < mpu.get_spiral_forward_virtual_size()
+                    else f"b{mpu.get_spiral_backward_virtual_size() - 1}"
+                )
+                prefetch_next = get_thunder_cuda_manager().Event(
+                    "prefetch",
+                    "compute",
+                    tag=tag,
+                    post_wait_fn=lambda fwd_stage_id=fwd_stage_id: set_module_spiral_status(
+                        model[fwd_stage_id + 1], SpiralParamStatus.GPU
+                    ),
+                )
+                if get_thunder_cuda_manager().record_event(prefetch_next) == -1:
+                    raise RuntimeError("record_event failed")
+                prefetch_event_queries[prefetch_next.tag] = prefetch_next
+
+        # compute stream wait for prefetch current stage
+        with torch.cuda.stream(get_thunder_cuda_manager().Stream("compute")):
+            if (
+                get_thunder_cuda_manager().wait_event(prefetch_event_queries.pop(f"prefetch:f{fwd_stage_id}"))
+                == -1
+            ):
+                raise RuntimeError("wait_event failed")
+
+        # fwd microbatches
+        for i in range(num_microbatches):
+            if _DEBUG_SCHEDULE:
+                spiral_print(f" microbatch {i}")
+            torch.cuda.nvtx.range_push(f"f[{fwd_stage_id}]m[{i}]")
+
+            # send input tensor ckpt
+            if not forward_only:
+                comm_input_ckpt(next(ckpt_send_recv_schedule), model, tensor_shape, dtype)
+
+            # set input tensor
+            if mpu.is_pipeline_first_stage():
+                input_tensor = None
+                recv_handles = None
+            else:
+                input_tensor, recv_handles = spiral_p2p.recv_input_tensor(
+                    tensor_shape,
+                    dtype,
+                    batch_p2p_comm=batch_p2p_comm,
+                    overlap_p2p_comm=overlap_p2p_comm,
+                    timers=timers,
+                )
+
+            with torch.cuda.stream(get_thunder_cuda_manager().Stream("compute")):
+                # wait for recv input tensor
+                # NOTE (SpiralPipe) Must be done in compute stream to avoid error
+                if recv_handles is not None:
+                    for req in recv_handles:
+                        req.wait()
+
+                # NOTE (SpiralPipe) Original FWD should not create computation graph for two reasons
+                # 1. It will be recomputed in BWD
+                # 2. It creates computation graph to the input ckpt tensor, which leads to computation graph duplication when BWD is performed in the same rank. Symptoms include size error due to ops with size 0 tensor.
+                with torch.no_grad():
+                    output_tensor = forward_step(
+                        forward_step_func,
+                        data_iterator[fwd_stage_id],
+                        model[fwd_stage_id],
+                        num_microbatches,
+                        input_tensor,
+                        forward_data_store,
+                        timers,
+                        collect_non_loss_data,
+                        dtype,
+                        enable_autocast,
+                    )
+
+                compute_microbatch = get_thunder_cuda_manager().Event(
+                    "compute",
+                    None,
+                    tag=f"compute:f{fwd_stage_id}m{i}",
+                )
+                if get_thunder_cuda_manager().record_event(compute_microbatch) == -1:
+                    raise RuntimeError("record_event failed")
+                compute_event_queries[compute_microbatch.tag] = compute_microbatch
+
+                if i == num_microbatches - 1:
+                    # notify free stream to act
+                    compute_microbatches_end = get_thunder_cuda_manager().Event(
+                        "compute", "free", tag=f"compute:f{fwd_stage_id}end"
+                    )
+                    if get_thunder_cuda_manager().record_event(compute_microbatches_end) == -1:
+                        raise RuntimeError("record_event failed")
+                    compute_event_queries[compute_microbatches_end.tag] = compute_microbatches_end
+
+            # send output tensor
+            if (
+                get_thunder_cuda_manager().wait_event(compute_event_queries.pop(f"compute:f{fwd_stage_id}m{i}"))
+                == -1
+            ):
+                raise RuntimeError("wait_event failed")
+            if not mpu.is_pipeline_last_stage():
+                if not mpu.get_pipeline_model_parallel_rank() == mpu.get_pipeline_model_parallel_world_size() - 1:
+                    # send output tensor immediately
+                    _ = spiral_p2p.send_output_tensor(
+                        output_tensor,
+                        overlap_p2p_comm=overlap_p2p_comm,
+                        batch_p2p_comm=batch_p2p_comm,
+                        timers=timers,
+                    )
+                else:
+                    # control path for only last worker
+                    # delay output tensor
+                    delayed_output_tensors.append(output_tensor)
+
+            if mpu.get_pipeline_model_parallel_rank() == mpu.get_pipeline_model_parallel_world_size() - 1:
+                if (curr_delay_output_tensors >= NUM_DELAY_OUTPUT_TENSORS):
+                    # send delayed output tensor
+                    if (len(delayed_output_tensors) > 0):
+                        _ = spiral_p2p.send_output_tensor(
+                            delayed_output_tensors.pop(0),
+                            overlap_p2p_comm=overlap_p2p_comm,
+                            batch_p2p_comm=batch_p2p_comm,
+                            timers=timers,
+                        )
+                else:
+                    curr_delay_output_tensors += 1
+            torch.cuda.nvtx.range_pop()
+        # end fwd microbatches
+
+        with torch.cuda.stream(get_thunder_cuda_manager().Stream("free")):
+            if (
+                get_thunder_cuda_manager().wait_event(compute_event_queries.pop(f"compute:f{fwd_stage_id}end"))
+                == -1
+            ):
+                raise RuntimeError("wait_event failed")
+
+            # free fwd stage
+            model[fwd_stage_id].spiral_free()
+            if not (
+                forward_only
+                and fwd_stage_id == mpu.get_spiral_forward_virtual_size() - 1
+            ):
+                free_curr = get_thunder_cuda_manager().Event(
+                    "free",
+                    "prefetch",
+                    tag=f"free:f{fwd_stage_id}",
+                )
+                if get_thunder_cuda_manager().record_event(free_curr) == -1:
+                    raise RuntimeError("record_event failed")
+                free_event_queries[free_curr.tag] = free_curr
+
+        mpu.set_spiral_forward_virtual_rank(None)
+
+    if forward_only:
+        _cleanup()
+        return forward_data_store
+    # end fwd
+
+    # bwd
+    assert not forward_only, "Forward only mode should have returned already"
+    for bwd_stage_id in range(
+        mpu.get_spiral_backward_virtual_size() - 1, -1, -1
+    ):
+        if _DEBUG_SCHEDULE:
+            spiral_print(f"Start bwd stage {bwd_stage_id}")
+        mpu.set_spiral_backward_virtual_rank(bwd_stage_id)
+
+        assert (
+            hasattr(model[-bwd_stage_id - 1], "spiral_backward_stage_id")
+            and getattr(model[-bwd_stage_id - 1], "spiral_backward_stage_id")
+            == bwd_stage_id
+        ), "Backward stage ID mismatch between virtual rank and model."
+
+        # prefetch next stage
+        with torch.cuda.stream(get_thunder_cuda_manager().Stream("prefetch")):
+            if (
+                get_thunder_cuda_manager().wait_event(
+                    free_event_queries.pop(
+                        f"free:f{fwd_stage_id}"
+                        if bwd_stage_id == mpu.get_spiral_backward_virtual_size() - 1
+                        else f"free:b{bwd_stage_id + 1}"
+                    )
+                )
+                == -1
+            ):
+                raise RuntimeError("wait_event failed")
+
+            if not bwd_stage_id == 0:
+                model[-bwd_stage_id].spiral_fetch(non_blocking=True)
+                prefetch_next = get_thunder_cuda_manager().Event(
+                    "prefetch",
+                    "compute",
+                    tag="prefetch:" + f"b{bwd_stage_id - 1}",
+                    post_wait_fn=lambda bwd_stage_id=bwd_stage_id: set_module_spiral_status(
+                        model[-bwd_stage_id], SpiralParamStatus.GPU
+                    ),
+                )
+                if get_thunder_cuda_manager().record_event(prefetch_next) == -1:
+                    raise RuntimeError("record_event failed")
+                prefetch_event_queries[prefetch_next.tag] = prefetch_next
+
+        # compute stream wait for prefetch current stage
+        # NOTE: prefetch for last bwd stage is called in the fwd loop
+        with torch.cuda.stream(get_thunder_cuda_manager().Stream("compute")):
+            if (
+                get_thunder_cuda_manager().wait_event(prefetch_event_queries.pop(f"prefetch:b{bwd_stage_id}"))
+                == -1
+            ):
+                raise RuntimeError("wait_event failed")
+
+        # bwd microbatches
+        for i in range(num_microbatches):
+            if _DEBUG_SCHEDULE:
+                spiral_print(f" microbatch {i}")
+            torch.cuda.nvtx.range_push(f"b[{bwd_stage_id}]m[{i}]")
+
+            # recv input tensor ckpt
+            recv_input_ckpts, recv_input_ckpt_handle = comm_input_ckpt(
+                next(ckpt_send_recv_schedule), model, tensor_shape, dtype
+            )
+
+            # set output tensor grad
+            if mpu.is_pipeline_last_stage():
+                output_tensor_grad = None
+                recv_handles = None
+            else:
+                output_tensor_grad, recv_handles = spiral_p2p.recv_output_tensor_grad(
+                    tensor_shape,
+                    dtype,
+                    batch_p2p_comm=batch_p2p_comm,
+                    overlap_p2p_comm=overlap_p2p_comm,
+                    timers=timers,
+                )
+
+            with torch.cuda.stream(get_thunder_cuda_manager().Stream("compute")):
+                # wait for recv input tensor ckpt
+                # NOTE (SpiralPipe) Must be done in compute stream to avoid error
+                if recv_input_ckpt_handle is not None:
+                    for req in recv_input_ckpt_handle:
+                        req.wait()
+                if recv_input_ckpts is not None:
+                    assert (
+                        len(recv_input_ckpts) == 1
+                    ), f"Only 1 input tensor ckpt expected. Got {len(recv_input_ckpts)}"
+                    input_tensor_ckpt = recv_input_ckpts.pop(0)
+
+                # wait for recv output tensor grad
+                # NOTE (SpiralPipe) Must be done in compute stream to avoid error
+                if recv_handles is not None:
+                    for req in recv_handles:
+                        req.wait()
+
+                output_tensor = forward_step(
+                    forward_step_func,
+                    data_iterator[
+                        bwd_stage_id + mpu.get_spiral_forward_virtual_size()
+                    ],
+                    model[-bwd_stage_id - 1],
+                    num_microbatches,
+                    input_tensor_ckpt,
+                    [],
+                    timers,
+                    collect_non_loss_data,
+                    dtype,
+                    enable_autocast,
+                )
+
+                if optimize_after_bwd_stage:
+                    # grad_scaler is aligned (ascending) w.r.t bwd_stage_id
+                    # (same as optimizer_list in SpiralStageOptimizer)
+                    _grad_scaler = grad_scaler[bwd_stage_id]
+                else:
+                    _grad_scaler = grad_scaler
+                input_tensor_grad = backward_step(
+                    _grad_scaler,
+                    input_tensor_ckpt,
+                    output_tensor,
+                    output_tensor_grad,
+                    model_type,
+                    timers,
+                    deallocate_pipeline_outputs,
+                )
+
+                compute_microbatch = get_thunder_cuda_manager().Event(
+                    "compute",
+                    None,
+                    tag=f"compute:b{bwd_stage_id}m{i}",
+                )
+                if get_thunder_cuda_manager().record_event(compute_microbatch) == -1:
+                    raise RuntimeError("record_event failed")
+                compute_event_queries[compute_microbatch.tag] = compute_microbatch
+
+                if i == num_microbatches - 1:
+                    # notify free stream to act
+                    compute_microbatches_end = get_thunder_cuda_manager().Event(
+                        "compute", None, tag=f"compute:b{bwd_stage_id}end"
+                    )
+                    if get_thunder_cuda_manager().record_event(compute_microbatches_end) == -1:
+                        raise RuntimeError("record_event failed")
+                    compute_event_queries[compute_microbatches_end.tag] = compute_microbatches_end
+
+            # send input tensor grad
+            if (
+                get_thunder_cuda_manager().wait_event(compute_event_queries.pop(f"compute:b{bwd_stage_id}m{i}"))
+                == -1
+            ):
+                raise RuntimeError("wait_event failed")
+            if not mpu.is_pipeline_first_stage():
+                if not mpu.get_pipeline_model_parallel_rank() == mpu.get_pipeline_model_parallel_world_size() - 1:
+                    # send input tensor grad immediately
+                    _ = spiral_p2p.send_input_tensor_grad(
+                        input_tensor_grad,
+                        overlap_p2p_comm=overlap_p2p_comm,
+                        batch_p2p_comm=batch_p2p_comm,
+                        timers=timers,
+                    )
+                else:
+                    # control path for only last worker
+                    # delay input tensor grad
+                    delayed_input_tensor_grads.append(input_tensor_grad)
+
+            if mpu.get_pipeline_model_parallel_rank() == mpu.get_pipeline_model_parallel_world_size() - 1:
+                if (curr_delay_input_tensor_grads >= NUM_DELAY_INPUT_TENSOR_GRADS):
+                    if (len(delayed_input_tensor_grads) > 0):
+                        _ = spiral_p2p.send_input_tensor_grad(
+                            delayed_input_tensor_grads.pop(0),
+                            overlap_p2p_comm=overlap_p2p_comm,
+                            batch_p2p_comm=batch_p2p_comm,
+                            timers=timers,
+                        )
+                else:
+                    curr_delay_input_tensor_grads += 1
+
+            torch.cuda.nvtx.range_pop()
+        # end bwd microbatches
+
+        with torch.cuda.stream(get_thunder_cuda_manager().Stream("free")):
+            if (
+                get_thunder_cuda_manager().wait_event(compute_event_queries.get(f"compute:b{bwd_stage_id}end"))
+                == -1
+            ):
+                raise RuntimeError("wait_event failed")
+
+            # free bwd stage
+            model[-bwd_stage_id - 1].spiral_free()
+            free_curr = get_thunder_cuda_manager().Event(
+                "free",
+                None if bwd_stage_id == 0 else "prefetch",
+                tag=f"free:b{bwd_stage_id}",
+            )
+            if get_thunder_cuda_manager().record_event(free_curr) == -1:
+                raise RuntimeError("record_event failed")
+            free_event_queries[free_curr.tag] = free_curr
+        # end free bwd stage
+
+        # if grad offload is not overlapped, then it should be reduced and offloaded after `forward_backward_func` at train_step()
+        if offload_grad_after_bwd_stage:
+            with torch.cuda.stream(get_thunder_cuda_manager().Stream("offload")):
+                if (
+                    get_thunder_cuda_manager().wait_event(compute_event_queries.pop(f"compute:b{bwd_stage_id}end"))
+                    == -1
+                ):
+                    raise RuntimeError("wait_event failed")
+                # if not spiral stage optimizer, then gradient should be manually reduced
+                if optimize_after_bwd_stage:
+                    optimizer[bwd_stage_id].reduce_model_grads(get_args(), get_timers())
+                else:
+                    model[-bwd_stage_id - 1].allreduce_gradients()
+                # offload grads
+                model[-bwd_stage_id - 1].spiral_offload_grad(non_blocking=True)
+                offload_grad_curr = get_thunder_cuda_manager().Event(
+                    "offload",
+                    "free",
+                    tag=f"offload_grad:b{bwd_stage_id}"
+                )
+                if get_thunder_cuda_manager().record_event(offload_grad_curr) == -1:
+                    raise RuntimeError("record_event failed")
+                offload_event_queries[offload_grad_curr.tag] = offload_grad_curr
+
+                # if not spiral stage optimizer, then optimizer will be executed after iteration
+                if optimize_after_bwd_stage:
+                    _offload_grad_ev_cpu = torch.cuda.Event() # event to synchronize at cpu
+                    _optimizer_thread_queue = Queue()
+                    _offload_grad_ev_cpu.record()
+
+                    inner_step_kwargs = {}
+                    inner_step_kwargs["spiral_offload_grad_ev"] = _offload_grad_ev_cpu
+                    inner_step_kwargs["spiral_optimizer_thread_queue"] = _optimizer_thread_queue
+                    inner_step_kwargs["spiral_offload_grad_ev_long"] = _offload_grad_ev_cpu.cuda_event
+
+                    # TODO (SpiralPipe) timers is None. Fix it
+                    op = threading.Thread(
+                        target=optimizer[bwd_stage_id].step,
+                        args=(get_args(), get_timers()),
+                        kwargs=inner_step_kwargs,
+                    )
+                    op.start()
+                    optimizer_threads.append((op, _optimizer_thread_queue))
+
+            with torch.cuda.stream(get_thunder_cuda_manager().Stream("free")):
+                if (
+                    get_thunder_cuda_manager().wait_event(offload_event_queries.pop(f"offload_grad:b{bwd_stage_id}"))
+                    == -1
+                ):
+                    raise RuntimeError("wait_event failed")
+
+                # free bwd stage grads
+                model[-bwd_stage_id - 1].spiral_free_grad()
+                free_grad_curr = get_thunder_cuda_manager().Event(
+                    "free",
+                    None,
+                    tag=f"free_grad:b{bwd_stage_id}",
+                )
+                if get_thunder_cuda_manager().record_event(free_grad_curr) == -1:
+                    raise RuntimeError("record_event failed")
+                free_event_queries[free_grad_curr.tag] = free_grad_curr
+        # end offload & free grad
+
+        mpu.set_spiral_backward_virtual_rank(None)
+    # end bwd
+
+    # do post-pipeline non-compute timesteps
+    assert not forward_only, "Forward only mode should have returned already"
+    __num_post_pipeline_non_compute_ts = (
+        mpu.get_pipeline_model_parallel_world_size()
+        - mpu.get_pipeline_model_parallel_rank()
+        - 1
+    )
+    for _ in range(__num_post_pipeline_non_compute_ts):
+        if not forward_only:
+            comm_input_ckpt(next(ckpt_send_recv_schedule), model, tensor_shape, dtype)
+
+    # cleanup schedule events
+    if (
+        get_thunder_cuda_manager().wait_event(free_event_queries.pop(f"free:b0"))
+        == -1
+    ):
+        raise RuntimeError("wait_event failed")
+
+    if offload_grad_after_bwd_stage:
+        for bwd_stage_id in range(mpu.get_spiral_backward_virtual_size() - 1, -1, -1):
+            # flush free grad event queries
+            if (
+                get_thunder_cuda_manager().wait_event(
+                    free_event_queries.pop(f"free_grad:b{bwd_stage_id}")
+                )
+                == -1
+            ):
+                raise RuntimeError("wait_event failed")
+
+    # join optimizer
+    if optimize_after_bwd_stage:
+        for op, q in optimizer_threads:
+            op.join()
+            kwargs["spiral_stage_optimizer_step_returns"].appendleft(q.get())
+
+    _cleanup()
+    return forward_data_store
